% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_absolute_moment}
\alias{compute_absolute_moment}
\title{Absolute Moment of Order k}
\usage{
compute_absolute_moment(obs, k)
}
\arguments{
\item{obs}{Observations (type: armadillo numeric vector).}

\item{k}{Moment order (type: double)}
}
\value{
Moment value (type: double)
}
\description{
Computes the absolute moment of order k of a sample of observations.
}
\examples{
rvs <- stabledist::rstable(100000, 1.9, 0.5, 1, 0, pm = 0)
absolute_moment <- compute_absolute_moment(rvs, 2)
}
