% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_chisq1_percentile}
\alias{get_chisq1_percentile}
\title{Chi^2(1) Percentile}
\usage{
get_chisq1_percentile(value)
}
\arguments{
\item{value}{Chi^2(1) value (type: double).}
}
\value{
Chi^2(1) percentile (type: double).
}
\description{
Returns the Chi^2(1) percentile for the test statistic.
}
\examples{
get_chisq1_percentile(20.0)
}
